+++ mypyc/class_ir.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Set, Tuple, Dict, Union, NamedTuple
 from collections import OrderedDict
 
 from mypyc.common import JsonDict
-from mypyc.ops import Value, DeserMaps
+from mypyc.ir.ops import Value, DeserMaps
 from mypyc.rtypes import RType, RInstance, deserialize_type
 from mypyc.func_ir import FuncIR, FuncDecl, FuncSignature
 from mypyc.namegen import NameGenerator, exported_name
