+++ mypyc/irbuild/context.py
@@ -2,7 +2,7 @@ from typing import List, Optional, Tuple
 
 from mypy.nodes import FuncItem
 
-from mypyc.ops import Value, BasicBlock, AssignmentTarget
+from mypyc.ir.ops import Value, BasicBlock, AssignmentTarget
 from mypyc.func_ir import INVALID_FUNC_DEF
 from mypyc.class_ir import ClassIR
 from mypyc.common import decorator_helper_name
