+++ mypyc/irbuild/mapper.py
@@ -8,7 +8,7 @@ from mypy.types import (
     get_proper_type
 )
 
-from mypyc.ops import LiteralsMap
+from mypyc.ir.ops import LiteralsMap
 from mypyc.rtypes import (
     RType, RUnion, RTuple, RInstance, object_rprimitive, dict_rprimitive, tuple_rprimitive,
     none_rprimitive, int_rprimitive, float_rprimitive, str_rprimitive, bool_rprimitive,
