+++ mypyc/irbuild/specialize.py
@@ -17,7 +17,7 @@ from typing import Callable, Optional, Dict, Tuple
 from mypy.nodes import CallExpr, RefExpr, MemberExpr, TupleExpr, GeneratorExpr, ARG_POS
 from mypy.types import AnyType, TypeOfAny
 
-from mypyc.ops import (
+from mypyc.ir.ops import (
     Value, BasicBlock, LoadInt, RaiseStandardError, Unreachable, OpDescription
 )
 from mypyc.rtypes import (
