+++ mypyc/irbuild/util.py
@@ -6,7 +6,7 @@ from mypy.nodes import (
     ARG_OPT, GDEF
 )
 
-from mypyc.ops import Environment, AssignmentTargetRegister
+from mypyc.ir.ops import Environment, AssignmentTargetRegister
 from mypyc.rtypes import RInstance
 from mypyc.class_ir import ClassIR
 from mypyc.common import SELF_NAME
