+++ mypyc/primitives/dict_ops.py
@@ -2,7 +2,7 @@
 
 from typing import List
 
-from mypyc.ops import EmitterInterface, ERR_FALSE, ERR_MAGIC, ERR_NEVER
+from mypyc.ir.ops import EmitterInterface, ERR_FALSE, ERR_MAGIC, ERR_NEVER
 from mypyc.rtypes import dict_rprimitive, object_rprimitive, bool_rprimitive, int_rprimitive
 
 from mypyc.primitives.registry import (
