+++ mypyc/primitives/list_ops.py
@@ -2,7 +2,7 @@
 
 from typing import List
 
-from mypyc.ops import ERR_MAGIC, ERR_NEVER, ERR_FALSE, EmitterInterface
+from mypyc.ir.ops import ERR_MAGIC, ERR_NEVER, ERR_FALSE, EmitterInterface
 from mypyc.rtypes import (
     int_rprimitive, short_int_rprimitive, list_rprimitive, object_rprimitive, bool_rprimitive
 )
