+++ mypyc/primitives/set_ops.py
@@ -3,7 +3,7 @@ from mypyc.primitives.registry import (
     func_op, method_op, binary_op,
     simple_emit, negative_int_emit, call_emit, call_negative_bool_emit,
 )
-from mypyc.ops import ERR_MAGIC, ERR_FALSE, ERR_NEVER, EmitterInterface
+from mypyc.ir.ops import ERR_MAGIC, ERR_FALSE, ERR_NEVER, EmitterInterface
 from mypyc.rtypes import object_rprimitive, bool_rprimitive, set_rprimitive, int_rprimitive
 from typing import List
