+++ mypyc/test/test_emit.py
@@ -3,7 +3,7 @@ import unittest
 from mypy.nodes import Var
 
 from mypyc.emit import Emitter, EmitterContext
-from mypyc.ops import BasicBlock, Environment
+from mypyc.ir.ops import BasicBlock, Environment
 from mypyc.rtypes import int_rprimitive
 from mypyc.namegen import NameGenerator
